+++ lib/src/components/general/day_separator/day_separator.dart
@@ -9,22 +9,15 @@ class DaySeparator extends StatelessWidget {
     required this.dayWidth,
   });
 
-  /// The number of days.
-  ///
-  /// This is used to determine the number of day separators to display.
   final int numberOfDays;
-
-  /// The width of a day. (rounded)
-  ///
-  /// This is if you want to display the day separators perfectly.
   final double dayWidth;
 
   @override
   Widget build(BuildContext context) {
     final style = CalendarStyleProvider.of(context).style.daySeparatorStyle;
 
-    final width = style?.thickness ?? 1;
-    final color = style?.color ?? Theme.of(context).colorScheme.surfaceVariant;
+    final width = style.thickness ?? 1;
+    final color = style.color ?? Theme.of(context).colorScheme.surfaceVariant;
     return Stack(
       fit: StackFit.expand,
       children: [