+++ lib/src/views/multi_day_view/multi_day_header.dart
@@ -44,6 +44,7 @@ class MultiDayHeader<T> extends StatelessWidget {
           }
 
           return Column(
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
               calendarHeader,
               if (viewConfiguration.showHeader) dayHeader,
@@ -191,7 +192,8 @@ class AnimatedMultiDayEventsHeader<T> extends StatelessWidget {
               duration: const Duration(milliseconds: 200),
               child: SizedBox(
                 height: viewConfiguration.multiDayTileHeight *
-                    (multiDayEventGroup.maxNumberOfStackedEvents + (viewConfiguration.createMultiDayEvents ? 1 : 0)),
+                    (multiDayEventGroup.maxNumberOfStackedEvents +
+                        (viewConfiguration.createMultiDayEvents ? 1 : 0)),
                 child: Stack(
                   clipBehavior: Clip.antiAlias,
                   children: [